microservice/package/requirements.txt
allennlp==2.8.0
allennlp_models==2.8.0
joblib==0.17.0
nltk==3.5
numpy==1.19.4
overrides
pandas==1.1.5
#pyspark==3.0.1
pyvis==0.1.9
requests==2.22.0
scikit_learn==1.0.2
seqeval==1.2.2
sklearn_crfsuite==0.3.6
spacy==2.3.7
importlib-metadata==4.13.*
#spark_nlp==2.6.3
#sparknlp==1.0.0
#neuralcoref==4.0
